wrap in in if in case nothing to go back

+++ src/molecules/molecule.js
@@ -367,24 +367,21 @@ export default class Molecule extends Atom {
    * Takes an array of recently deleted atoms
    */
   undo() {
-    let rawFile = JSON.parse(
-      GlobalVariables.recentMoleculeRepresentation.pop()
-    );
+    if (GlobalVariables.recentMoleculeRepresentation.length > 0) {
+      let rawFile = JSON.parse(
+        GlobalVariables.recentMoleculeRepresentation.pop()
+      );
+      const nodesCopy = [...GlobalVariables.topLevelMolecule.nodesOnTheScreen];
+      // Delete nodes so deserialize doesn't repeat, could be useful to not delete for a diff in the future
+      nodesCopy.forEach((atom, index) => {
+        atom.deleteNode();
+      });
 
-    // Create a shallow copy of nodesOnTheScreen using the spread operator
-    const nodesCopy = [...GlobalVariables.topLevelMolecule.nodesOnTheScreen];
-
-    // Delete nodes so deserialize doesn't repeat, could be useful to not delete for a diff in the future
-    nodesCopy.forEach((atom, index) => {
-      atom.deleteNode();
-    });
-
-    if (rawFile.fileTypeVersion == 1) {
-      GlobalVariables.topLevelMolecule.deserialize(rawFile);
+      if (rawFile.fileTypeVersion == 1) {
+        GlobalVariables.topLevelMolecule.deserialize(rawFile);
+      }
+      GlobalVariables.currentMolecule.selected = true;
     }
-    GlobalVariables.currentMolecule.selected = true;
-    console.log(GlobalVariables.recentMoleculeRepresentation);
-    //GlobalVariables.currentMolecule.placeAtom(item, true);
   }
 
   /**
